fixed opening animation
