+++ lib/workouts/widgets/widgets.dart
@@ -1,7 +1,7 @@
 export 'exercise_card/exercise_card.dart';
 export 'exercise_card/exercise_card_data.dart';
 export 'exercise_card/exercise_set_data.dart';
-
 export 'workout_card.dart';
+export 'workout_creator_body.dart';
 export 'workout_date_chip.dart';
 export 'workout_stats.dart';
